+++ shared/GR_Dataset.lua
@@ -19,8 +19,11 @@ function ds:WhatsNew() -- What's new in the current version
                     Discord: https://discord.gg/ZtS6Q2sKRH
                 (or click on the icon in the top left corner)|r
 
-    |CFFFFFF00v3.0.37 Notes|r
+    |CFFFFFF00v3.1.40 Notes|r
         - Added option to keep addon running and ignore certain ways to close it.
+        - esMX (Spanish Mexico) localization added.
+        - Fixed issue with whispers not being sent when player joins guild.
+        - Fixed issue with compact mode not removing anti-spam/black list before adding to invite list.
     |CFFFFFF00v3.0.36 Notes|r  -- Joined forces with FGI to bring recruiting to the next level.
         - Changed the 6 months to 180 days.
         (NOTE: If you were using the 6 months, you will need to update your settings.)